test correction 100

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -43,6 +43,6 @@ int _atoi(char *s)
 		i++;
 	}
 	if (entered)
-		return (digit * sign);
+		return ((int unsigned)digit * sign);
 	return (0);
 }
